+++ src/test/java/seedu/duke/ui/AsciiArtTest.java
@@ -1,0 +1,7 @@
+package seedu.duke.ui;
+
+import org.junit.jupiter.api.Test;
+
+public class AsciiArtTest {
+
+}
