+++ src/test/java/seedu/duke/ui/FormatterTest.java
@@ -1,0 +1,4 @@
+package seedu.duke.ui;
+
+public class FormatterTest {
+}
